ai-service/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
sentencepiece
rasterio 
aiohttp 
shapely 
numpy
pillow
redis
